Refactor modules and update docstrings
