+++ zero_liftsim/__init__.py
@@ -1,8 +1,19 @@
 """Top level package for zero_liftsim."""
 
-__all__ = ["__version__", "main", "state_viz"]
-__version__ = "0.1.0"
-
-from . import main
+from .simulation import Simulation, run_alpha_sim
+from .events import Event, ArrivalEvent, BoardingEvent, ReturnEvent
 from . import state_viz
 from . import pandas_ext  # noqa: F401 -- register pandas accessor
+
+__version__ = "0.1.0"
+
+__all__ = [
+    "__version__",
+    "Simulation",
+    "run_alpha_sim",
+    "Event",
+    "ArrivalEvent",
+    "BoardingEvent",
+    "ReturnEvent",
+    "state_viz",
+]
