+++ zero_liftsim/logging.py
@@ -1,3 +1,5 @@
+"""Simple file-based event logger."""
+
 from __future__ import annotations
 
 import json
